+++ leiribook/routes/web.php
@@ -38,12 +38,14 @@ Route::get('/books/filter', [LivroController::class, 'filterByGenre'])->name('fi
 
 Auth::routes(['register' => true, 'verify' => true]);
 Route::group(['middleware' => ['auth', 'verified']], function () {
+    Route::get('/profile/{user}', [UserController::class, 'profile'])->name('profile');
     route::get('editpassword', [UserController::class, 'editpassword'])->name('editpassword');
     route::patch('updatepassword', [UserController::class, 'updatepassword'])->name('updatepassword');
     Route::get('/users/{user}/send_reactivate_mail', [UserController::class, 'send_reactivate_email'])->name('users.sendActivationEmail');
+    Route::delete('/profile/{user}/destroy_photo', [PageController::class, 'destroy_photo_profile'])->name('profile.destroyPhotoProfile');
     route::group(['middleware' => ['role'], 'prefix' => 'admin', 'as' => 'admin.'], function () {
         Route::get('/', [PageController::class, 'admin'])->name('dashboard');
-        Route::delete('/users/{user}/destroy_photo', [UserController::class, 'destroy_photo'])->name('users.destroyPhoto');
+        Route::delete('admin/users/{user}/destroy_photo', [UserController::class, 'destroy_photo'])->name('admin.users.destroyPhoto');
         Route::resource('evento', EventoController::class);
         Route::resource('faqs', FaqController::class);
         Route::resource('users', UserController::class);
